(Spring 4): * Fix rutas, * Fix middleware, * Se cargaron 2 productos nuevos por formulario, * Cambios menores
