+++ public/css/normalize.css
@@ -163,10 +163,6 @@
 }
 
 
-@font-face { 
-    font-family: Bitsumishi; src: url(desing/fonts/BITSUMIS.TTF); }
-
-
 /* General clear*/
 *{
     margin: 0px;
