+++ src/routes/webRoutes.js
@@ -1,21 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const multer = require('multer');
 const path = require('path');
 const webControllers = require('../controllers/webControllers');
 
-// ---- Multer ---- //
-let storage = multer.diskStorage( {
-    destination: (req, file, callback) => {
-        let folder = path.join(__dirname, '../../public/images/products');
-        callback(null, folder);
-    },
-    filename: function(req, file, callback) {
-        let imageName = 'product-' + Date.now() + path.extname(file.originalname);
-        callback(null, imageName);
-    }
-});
-let fileUpload = multer({ storage });
 
 router.get('/', webControllers.index);
