+++ src/userLogs.txt
@@ -8,3 +8,38 @@ El usuario ingreso a: /products/detail
 El usuario ingreso a: /products/detail
 El usuario ingreso a: /products/detail
 El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /login
+El usuario ingreso a: /register
+El usuario ingreso a: /
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productCreateForm
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /products/productCreateForm
+El usuario ingreso a: /products/productsList
+El usuario ingreso a: /
+El usuario ingreso a: /products/productDetail
+El usuario ingreso a: /
