+++ src/views/partials/head.ejs
@@ -1,32 +1,37 @@
-<meta charset="UTF-8">
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Tecnocom</title>
-<link rel="shortcut icon" href="desing/tcom.ico">
-<link rel="stylesheet" href="css/normalize.css">
-<link rel="stylesheet" href="css/index-styles.css">
-<link rel="stylesheet" href="css/register-styles.css">
-<link rel="stylesheet" href="css/header-styles.css">
-<link rel="stylesheet" href="css/product-styles.css">
-<link rel="stylesheet" href="css/footer-styles.css">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Tecnocom</title>
+    <link rel="stylesheet" href="/css/normalize.css">
+    <link rel="stylesheet" href="/css/index-styles.css">
+    <link rel="stylesheet" href="/css/register-styles.css">
+    <link rel="stylesheet" href="/css/header-styles.css">
+    <link rel="stylesheet" href="/css/product-styles.css">
+    <link rel="stylesheet" href="/css/footer-styles.css">
         
-<style> @font-face { 
-        font-family: BauerRegular; src: url(desing/fonts/BAUERG__.TTF); }
-    @font-face { 
-        font-family: OsakaSansSerif; src: url(desing/fonts/osaka-re.ttf); }
-    @font-face { 
-        font-family: ElgethyBoldOblique; src: url(desing/fonts/Elgethy_Bold_Oblique.ttf); }
-    @font-face { 
-        font-family: HandSean; src: url(desing/fonts/handsean.ttf); }
-    @font-face { 
-        font-family: DanielBK; src: url(desing/fonts/danielbk.ttf); }
-    @font-face { 
-        font-family: Coolvetica; src: url(desing/fonts/coolvetica_rg.ttf); }
-    @font-face { 
-        font-family: Jadefedgah; src: url(desing/fonts/Jadefedgah[80].ttf); }
-</style>
+    <style>
+        @font-face { 
+            font-family: BauerRegular; src: url(/desing/fonts/BAUERG__.TTF); }
+        @font-face { 
+            font-family: OsakaSansSerif; src: url(/desing/fonts/osaka-re.ttf); }
+        @font-face { 
+            font-family: Bitsumishi; src: url(/desing/fonts/bitsumis.ttf); }
+        @font-face { 
+            font-family: ElgethyBoldOblique; src: url(/desing/fonts/Elgethy_Bold_Oblique.ttf); }
+        @font-face { 
+            font-family: HandSean; src: url(/desing/fonts/handsean.ttf); }
+        @font-face { 
+            font-family: DanielBK; src: url(/desing/fonts/danielbk.ttf); }
+        @font-face { 
+            font-family: Coolvetica; src: url(/desing/fonts/coolvetica_rg.ttf); }
+        @font-face { 
+            font-family: Jadefedgah; src: url(/desing/fonts/Jadefedgah[80].ttf); }
+    </style>
 
-<link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
-<link href="/desing/fontawesome6/css/fontawesome.css" rel="stylesheet">
-<link href="/desing/fontawesome6/css/brands.css" rel="stylesheet">
-<link href="/desing/fontawesome6/css/solid.css" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
+    <link href="/desing/fontawesome6/css/fontawesome.css" rel="stylesheet">
+    <link href="/desing/fontawesome6/css/brands.css" rel="stylesheet">
+    <link href="/desing/fontawesome6/css/solid.css" rel="stylesheet">
+    <link rel="shortcut icon" href="/desing/logo/tcom.ico">
+</head>
