+++ src/views/products/productsList.ejs
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+    <!-- Head -->
+    <%- include('../partials/head.ejs', {title: 'Ver todo '}) %>
+    <!-- /Head -->
+
+    <body>
+        <!-- Header -->
+        <%- include('../partials/header.ejs') %>
+        <!-- /Header -->
+
+        <main class="main-index">
+            <div class="linea"></div>
+            <h3 class="titulo-main-index">Página vacía para pruebas</h3>
+            <div class="bar-ccb" id="cc-bar"></div>
+            <div class="linea"></div>
+
+
+        </main>
+        
+        <!-- Footer -->
+        <%- include('../partials/footer.ejs') %>
+        <!-- /Footer -->
+    </body>
+</html>
