+++ src/views/web/login.ejs
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
 <html lang="es">
-  <head>
-    <!-- head -->
-    <%- include('../partials/head.ejs') %>
-  </head>
+  <!-- Head -->
+  <%- include('../partials/head.ejs') %>
+  <!-- /Head -->
 
     <body>
         <!-- header -->
