+++ src/views/web/register.ejs
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
 <html lang="es">
-    <head>
-        <!-- head -->
-        <%- include('../partials/head.ejs') %>
-    </head>
+    <!-- Head -->
+    <%- include('../partials/head.ejs') %>
+    <!-- /Head -->
+    
     <body>
         <!-- header -->
         <%- include('../partials/header.ejs') %>
